Rename TaxonPresenter to TaxonPayloadBuilder
The class is responsible to generate a payload, therefore the name Presenter
could be confusing. By explicitly calling it TaxonPayloadBuilder we make its
purpose clearer.
